requirements-windows.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
blinker==1.6.2
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.1
cryptography==41.0.4
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.4
django-baton==2.8.0
django-braces==1.15.0
django-debug-toolbar==4.2.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.1.0
django-stubs==4.2.4
django-stubs-ext==4.2.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
easy-thumbnails==2.8.5
factory-boy==3.3.0
Faker==19.6.2
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
kombu==5.3.2
MarkupSafe==2.1.3
model-bakery==1.15.0
mypy==1.5.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.1
pipdeptree==2.13.0
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-magic-bin
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2018.5
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
rsa==4.9
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
text-unidecode==1.3
types-pytz==2023.3.0.1
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
